added files to support x86-64 codegen from instructions refactored instruction operand payloads to be 32 bits each. as this seems like more than enough for most compilations, and we can remove some of the overhead of dynamically allocating each individual local variable, instead keeping all that data in a single dynamic array. which should improve cache locality. (using indices instead of pointers is more stable, or if you like less coupled)

+++ exp/include/env/string_interner.h
@@ -27,8 +27,8 @@
  * ptr comparison
  */
 typedef struct StringInterner {
-    u64 capacity;
-    u64 count;
+    u32 capacity;
+    u32 count;
     ConstantString **buffer;
 } StringInterner;
